requirements.txt
python-telegram-bot==13.15
requests==2.26.0